Improve messaging/calendar create contact performance
