+++ packages/twenty-server/src/modules/connected-account/auto-companies-and-contacts-creation/create-contact/create-contact.service.ts
@@ -7,6 +7,7 @@ import { PersonRepository } from 'src/modules/person/repositories/person.reposit
 import { getFirstNameAndLastNameFromHandleAndDisplayName } from 'src/modules/calendar-messaging-participant/utils/get-first-name-and-last-name-from-handle-and-display-name.util';
 import { InjectObjectMetadataRepository } from 'src/engine/object-metadata-repository/object-metadata-repository.decorator';
 import { PersonObjectMetadata } from 'src/modules/person/standard-objects/person.object-metadata';
+import { ObjectRecord } from 'src/engine/workspace-manager/workspace-sync-metadata/types/object-record';
 
 type ContactToCreate = {
   handle: string;
@@ -50,16 +51,16 @@ export class CreateContactService {
     });
   }
 
-  public async createContacts(
+  public async createPeople(
     contactsToCreate: ContactToCreate[],
     workspaceId: string,
     transactionManager?: EntityManager,
-  ): Promise<void> {
-    if (contactsToCreate.length === 0) return;
+  ): Promise<ObjectRecord<PersonObjectMetadata>[]> {
+    if (contactsToCreate.length === 0) return [];
 
     const formattedContacts = this.formatContacts(contactsToCreate);
 
-    await this.personRepository.createPeople(
+    return await this.personRepository.createPeople(
       formattedContacts,
       workspaceId,
       transactionManager,
