+++ packages/twenty-server/src/modules/connected-account/auto-companies-and-contacts-creation/services/create-company-and-contact.service.ts
@@ -20,6 +20,7 @@ import { MessageParticipantService } from 'src/modules/messaging/services/messag
 import { CalendarEventParticipantService } from 'src/modules/calendar/services/calendar-event-participant/calendar-event-participant.service';
 import { filterOutContactsFromCompanyOrWorkspace } from 'src/modules/connected-account/auto-companies-and-contacts-creation/utils/filter-out-contacts-from-company-or-workspace.util';
 import { FeatureFlagEntity } from 'src/engine/core-modules/feature-flag/feature-flag.entity';
+import { ObjectRecord } from 'src/engine/workspace-manager/workspace-sync-metadata/types/object-record';
 
 @Injectable()
 export class CreateCompanyAndContactService {
@@ -37,14 +38,14 @@ export class CreateCompanyAndContactService {
     private readonly featureFlagRepository: Repository<FeatureFlagEntity>,
   ) {}
 
-  async createCompaniesAndContacts(
+  async createCompaniesAndPeople(
     connectedAccountHandle: string,
     contactsToCreate: Contacts,
     workspaceId: string,
     transactionManager?: EntityManager,
-  ) {
+  ): Promise<ObjectRecord<PersonObjectMetadata>[]> {
     if (!contactsToCreate || contactsToCreate.length === 0) {
-      return;
+      return [];
     }
 
     // TODO: This is a feature that may be implemented in the future
@@ -68,7 +69,7 @@ export class CreateCompanyAndContactService {
     );
 
     if (uniqueHandles.length === 0) {
-      return;
+      return [];
     }
 
     const alreadyCreatedContacts = await this.personRepository.getByEmails(
@@ -120,7 +121,7 @@ export class CreateCompanyAndContactService {
             : undefined,
       }));
 
-    await this.createContactService.createContacts(
+    return await this.createContactService.createPeople(
       formattedContactsToCreate,
       workspaceId,
       transactionManager,
@@ -139,7 +140,7 @@ export class CreateCompanyAndContactService {
 
     await workspaceDataSource?.transaction(
       async (transactionManager: EntityManager) => {
-        await this.createCompaniesAndContacts(
+        const createdPeople = await this.createCompaniesAndPeople(
           connectedAccountHandle,
           contactsToCreate,
           workspaceId,
@@ -147,11 +148,13 @@ export class CreateCompanyAndContactService {
         );
 
         await this.messageParticipantService.updateMessageParticipantsAfterPeopleCreation(
+          createdPeople,
           workspaceId,
           transactionManager,
         );
 
         await this.calendarEventParticipantService.updateCalendarEventParticipantsAfterPeopleCreation(
+          createdPeople,
           workspaceId,
           transactionManager,
         );
